Assigns redundantOpCount when opening the cache.
Fixes a bug where redundantOpCount was not being tracked when
reading the journal file. This causes the journal file to grow
unless there are enough redundant ops while the cache is live.

In order to properly ensure that the cache is trimmed to size
between opening and closing the cache, redundantOpCount needs to
be counted when the cache is opened.